Use one DB pool for entire server
Previously we created a DB pool on every call to a function in the `db`
module.  This caused connection exhaustion (see #36).

Instead, introduce a new `struct KbsDb` which holds the pool. All the
methods in the `db` module are now methods of the struct. It is
initialized only once in `create_service()`, and then passed on to
whomever needs DB access. That part caused a lot of little changes.

Also add a `db::tests::test_stress` which performs queries in a loop
(this would fail with the earlier implementation because we query would
create a new DB pool).

Fixes #36

Signed-off-by: Dov Murik <dov.murik1@il.ibm.com>